orchestrator-agent/pyproject.toml
[tool.poetry]
name = "orchestrator-agent"
version = "0.1.0"
description = "Orchestrator agent for coordinating data analysis workflows"
authors = ["Google <no-reply@google.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "orchestrator_agent"}]

[tool.poetry.dependencies]
python = "^3.9"
google-adk = "*"
a2a-sdk = ">=0.2.12"
fastapi = ">=0.115.0"
uvicorn = ">=0.34.0"
httpx = ">=0.28.0"
pydantic = ">=2.10.0"
python-dotenv = ">=1.0.1"
APScheduler = ">=3.10.4"
opentelemetry-api = ">=1.21.0"
opentelemetry-sdk = ">=1.21.0"
opentelemetry-exporter-jaeger-thrift = ">=1.21.0"
opentelemetry-exporter-prometheus = ">=1.12.0rc1"
opentelemetry-instrumentation-httpx = ">=0.42b0"
opentelemetry-instrumentation-fastapi = ">=0.42b0"
psutil = ">=5.9.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
google-adk = { version = "^1.0.0", extras = ["eval"] }
pytest-asyncio = "^0.26.0"
psutil = "^5.9.0"
requests = "^2.31.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"
cloudpickle = "^3.1.1"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
